+++ consensus/ibft/src/main/java/org/hyperledger/besu/consensus/ibft/messagewrappers/RoundChange.java
@@ -103,8 +103,7 @@ public class RoundChange extends BftMessage<RoundChangePayload> {
     } else {
       rlpOut.writeNull();
     }
-    // Omit blockAccessList when absent so pre-26.1.0 peers can decode this message.
-    blockAccessList.ifPresent(bal -> bal.writeTo(rlpOut));
+    blockAccessList.ifPresentOrElse((bal) -> bal.writeTo(rlpOut), rlpOut::writeNull);
     rlpOut.endList();
     return rlpOut.encoded();
   }